% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{eval.quad}
\alias{eval.quad}
\title{Evaluation of multivariate normal distributed expectations}
\usage{
eval.quad(FUN = function(x) 1, X = NULL, ..., W = NULL, debug = FALSE)
}
\arguments{
\item{FUN}{log likelihood function of the parameters to be estimated.
Defaults to \code{funtion(x) 1}, in which case only the built-in multivariate normal pdf is evaluated.}

\item{X}{Matrix of quadrature points, see \code{\link{init.quad}}. Alternatively, the list of quadrature points and weights produced by \code{\link{init.quad}}.}

\item{...}{Additional arguments passed on to FUN.}

\item{W}{Vector of weights, or \code{NULL} if provided by \code{X}.}

\item{debug}{Logical, should we return evaluated FUN results?}
}
\value{
A vector with the evaluated integrals, as well as a covariance matrix attribute.
}
\description{
Evaluates the expectation of a (set of) parameters with a given function under a (built-in) multivariate normal distribution by Gauss-Hermite quadrature.
}
\details{
The evaluated function is assumed to have a multivariate normal distribution, with a given mean vector and covariance matrix.
The default identity function \code{FUN(x) = 1} reduces a multivariate normal distribution with mean vector \code{mu} and covariance matrix \code{Sigma}.

The integral under evaluation is;
\deqn{\int_{-\Infty}^{\Infty} N(\mu, \Sigma) \times FUN(X) \times X dX}{Int Prior * FUN(X) * X dX}

Hence, if left default, the result is the expectation \eqn{E(N(\mu,\Sigma))}{ of N(mu, Sigma)}.

Note: FUN is evaluated in a loop, vectorisation is a future possibility. FUN must return a single scalar on the log-scale.
}
\seealso{
\code{\link{init.quad}} for creating quadrature points.
}

