% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package-docs.R
\docType{package}
\name{MultiGHQuad-package}
\alias{MultiGHQuad,}
\alias{MultiGHQuad-package}
\title{Multidimensional Gauss-Hermite Quadrature}
\description{
Functions to perform n-dimensional numerical integration on n parameters with a multivariate normal prior distribution.
}
\details{
Use \code{\link{init.quad}} to generate a quadrature grid, and \code{\link{eval.quad}} to evaluate the integral.
}
\author{
Karel A Kroeze, \email{k.a.kroeze@utwente.nl}
}
\seealso{
init.quad, eval.quad
}

