% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{eval.quad}
\alias{eval.quad}
\title{Evaluation of multivariate normal distributed integral}
\usage{
eval.quad(FUN = function(x) 1, X = NULL, W = NULL, log = FALSE, ...)
}
\arguments{
\item{FUN}{(Likelihood) function of the parameters to be estimated.
Defaults to \code{funtion(x) 1}, in which case only the built-in multivariate normal pdf is evaluated.}

\item{X}{Matrix of quadrature points, see \code{\link{MGHQuadPoints}}. Alternatively, the list of quadrature points and weights produced by \code{\link{MGHQuadPoints}}.}

\item{W}{Vector of weights, or \code{NULL} if provided by \code{X}.}

\item{log}{Should we take the logarithm of FUN values? Defaults to FALSE, assuming FUN returns a Log-likelihood.}

\item{...}{Additional arguments passed on to FUN.}
}
\value{
A vector with the evaluated integrals, as well as a covariance matrix attribute.
}
\description{
Evaluates a given function with a (built-in) multivariate normal distribution with Gauss-Hermite quadrature.
}
\details{
The evaluated function is assumed to have a multivariate normal distribution, with a given mean vector and covariance matrix.
The default identity function \code{function(x) 1} reduces to an integral over a multivariate normal distribution with mean vector \code{mu} and covariance matrix \code{Sigma}.
}
\examples{
quadPoints <- MGHQuadPoints(Q=3)
# expected value of 3-dimensional multivariate normal distribution: N(0,1).
# (Since mean is currently fixed at zero, this is always zero.)
integral <- MGHQuadEval(Q=3,X=quadPoints)
integral
round(integral)
}
\seealso{
\code{\link{MGHQuadPoints}} for creating quadrature points.
}

